member and admin access restrictions implemented

+++ chor/views.py
@@ -1,16 +1,18 @@
 
+from cmdjango.views import render_chor
 from .models import Chor, Song, SongPropertyValue, SongPropertyName, SongPerformance
 from .forms import SongPropertyNameForm, SongPerformanceForm
 
 from django.db.models import Count, Q
+from django.contrib import messages
 from django.contrib.auth.decorators import login_required
 from django.core.handlers.wsgi import WSGIRequest
-from django.shortcuts import render, redirect
+from django.shortcuts import render, redirect, get_object_or_404
 
-from .views_Chor_CRUD import *
-from .views_Song_CRUD import *
-from .views_members import *
-from .views_InviteLinks import *
+from .views_Chor_CRUD import createChor, deleteChor, updateChor
+from .views_Song_CRUD import song, createSong, deleteSong, updateSong
+from .views_members import chorMembers, kickMember, makeAdmin, makeMember
+from .views_InviteLinks import createInviteLink, inviteLinkAccessed
 
 # General chor views
 
@@ -18,16 +20,28 @@ from .views_InviteLinks import *
 @login_required(login_url='user-login')
 def chorhomepage(request: WSGIRequest, chor_id):
     chor = get_object_or_404(Chor, id=chor_id)
+
+    # access restriction
+    if not chor.user_is_member(request.user.pk):
+        messages.error(request, 'You are not a member of the choir')
+        return redirect('user-homepage')
+    
     context = {
         'chor': chor,
         'backlink': '/',
     }
-    return render(request, 'chor/chor-homepage.html', context)
+    return render_chor(request, 'chor/chor-homepage.html', context)
 
 
 @login_required(login_url='user-login')
 def chorsongs(request: WSGIRequest, chor_id):
     chor = Chor.objects.get(id=chor_id)
+
+    # access restriction
+    if not chor.user_is_member(request.user.pk):
+        messages.error(request, 'You are not a member of the choir')
+        return redirect('user-homepage')
+
     sortby = request.GET.get('s')
     qstr = request.GET.get('q')
     if sortby:
@@ -71,12 +85,18 @@ def chorsongs(request: WSGIRequest, chor_id):
         'qstr': qstr,
         'backlink': "../"
     }
-    return render(request, 'chor/chor-songs.html', context)
+    return render_chor(request, 'chor/chor-songs.html', context)
 
 
 @login_required(login_url='user-login')
 def chorPerformances(request: WSGIRequest, chor_id):
     chor = Chor.objects.get(id=chor_id)
+
+    # access restriction
+    if not chor.user_is_member(request.user.pk):
+        messages.error(request, 'You are not a member of the choir')
+        return redirect('user-homepage')
+
     performances = SongPerformance.objects.filter(song__chor=chor)
     out_list = list()
     if len(performances):
@@ -91,53 +111,38 @@ def chorPerformances(request: WSGIRequest, chor_id):
         'performances': out_list,
         'backlink': '../',
     }
-    return render(request, 'chor/chor-performances.html', context)
+    return render_chor(request, 'chor/chor-performances.html', context)
 
 
 @login_required(login_url='user-login')
 def chorSongProperties(request: WSGIRequest, chor_id):
     chor = Chor.objects.get(id=chor_id)
+
+    # access restriction
+    if not chor.user_is_member(request.user.pk):
+        messages.error(request, 'You are not a member of the choir')
+        return redirect('user-homepage')
+
     properties = chor.songpropertyname_set.all()
+
     context = {
         'chor': chor,
         'properties': properties,
         'backlink': f'/chor{chor.pk}/songs/',
     }
-    return render(request, 'chor/chor-songproperties.html', context)
-
-
-# Song pages
-@login_required(login_url='user-login')
-def song(request: WSGIRequest, pk):
-    song = Song.objects.get(id=pk)
-    songpropertynames = song.chor.songpropertyname_set.all()
-
-    ''' Initializing attributes dict with empty values of all fields '''
-    attributes = dict()
-    for name in songpropertynames:
-        attributes.update({name.name: ''})
-    songpropvalues = song.songpropertyvalue_set.all()
-
-    ''' Filling attributes dict with proopertyvalues from database '''
-    for spv in songpropvalues:
-        attributes.update({spv.songpropertyname.name: spv.value})
-
-    ''' Finding performances of the song '''
-    performances = song.songperformance_set.all()
-
-    ''' Putting out '''
-    context = {
-        'song': song,
-        'attributes': attributes,
-        'performances': performances,
-        'backlink': f'/chor{song.chor.pk}/songs/'
-    }
-    return render(request, 'chor/song.html', context)
+    return render_chor(request, 'chor/chor-songproperties.html', context)
 
 
 # SongProperty views
 @login_required(login_url='user-login')
 def createProperty(request: WSGIRequest, chor_id):
+
+    # access restriction
+    chor = Chor.objects.get(id=chor_id)
+    if not chor.user_is_admin(request.user.pk):
+        messages.error(request, 'Admin level access needed')
+        return redirect('chor-songs', chor.pk)
+
     form = SongPropertyNameForm()
     if request.method == "POST":
         form = SongPropertyNameForm(request.POST)
@@ -155,10 +160,16 @@ def createProperty(request: WSGIRequest, chor_id):
 @login_required(login_url='user-login')
 def songPropertyNameDelete(request: WSGIRequest, prop_id):
     prop = SongPropertyName.objects.get(id=prop_id)
-    chor_id = prop.chor.pk
+    chor = prop.chor
+
+    # access restriction
+    if not prop.chor.user_is_admin(request.user.pk):
+        messages.error(request, 'Admin level access needed')
+        return redirect('chor-properties', chor.pk)
+
     if request.method == "POST":
         prop.delete()
-        return redirect('chor-songproperties', chor_id=chor_id)
+        return redirect('chor-songproperties', chor_id=chor.pk)
     context = {
         'obj': prop,
     }
@@ -169,10 +180,16 @@ def songPropertyNameDelete(request: WSGIRequest, prop_id):
 @login_required(login_url='user-login')
 def deletePerformance(request: WSGIRequest, perf_id):
     perf = SongPerformance.objects.get(id=perf_id)
-    chor_id = perf.song.chor.pk
+    chor = perf.song.chor
+
+    # access restriction
+    if not chor.user_is_admin(request.user.pk):
+        messages.error(request, 'Admin level access needed')
+        return redirect('song', perf.song.pk)
+
     if request.method == "POST":
         perf.delete()
-        return redirect('chor-performances', chor_id=chor_id)
+        return redirect('chor-performances', chor_id=chor.pk)
     context = {
         'obj': perf,
         'backlink': f'/song/{perf.song.pk}'
@@ -183,6 +200,12 @@ def deletePerformance(request: WSGIRequest, perf_id):
 @login_required(login_url='user-login')
 def createPerformance(request: WSGIRequest, song_id):
     song = Song.objects.get(id=song_id)
+
+    # access restriction
+    if not song.chor.user_is_admin(request.user.pk):
+        messages.error(request, 'Admin level access needed')
+        return redirect('song', song.pk)
+
     form = SongPerformanceForm()
     if request.method == "POST":
         form = SongPerformanceForm(request.POST)
